refactor package.json scripts
