+++ tests/testing-utils/index.ts
@@ -1,8 +1,6 @@
-import { exec } from 'node:child_process'
 // @ts-ignore
 import { spinNetworkUp, spinNetworkDown, } from "@entropyxyz/sdk/testing"
 import * as readline from 'readline'
-import { randomBytes } from 'crypto'
 
 export {
   spinNetworkUp,
@@ -10,63 +8,14 @@ export {
 }
 
 export * from './constants.mjs'
+export * from './promise-runner.mjs'
+export * from './exec-promise.mjs'
 export * from './setup-test'
-
-/* Promise wrapper function of [exec](https://nodejs.org/api/child_process.html#child_processexeccommand-options-callback)
- *
- * @param {string} command - a string command to run in child process
- */
-
-export function execPromise (command: string): Promise<any> {
-  return new Promise((res, rej) => {
-    exec(command, (error, stdout, stderr) => {
-      if (!error && !stderr) res(stdout)
-      else if (!!stderr && !error) rej(stderr)
-      else if (!!error) rej(error)
-    })
-  })
-}
-
-/* Helper for wrapping promises which makes it super clear in logging if the promise
- * resolves or threw.
- *
- * @param {any} t - an instance to tape runner
- * @param {boolean} keepThrowing - toggle throwing
- */
-export function promiseRunner(t: any, keepThrowing = false) {
-  // NOTE: this function swallows errors
-  return async function run(
-    message: string,
-    promise: Promise<any>
-  ): Promise<any> {
-    if (promise.constructor !== Promise) {
-      t.pass(message)
-      return Promise.resolve(promise)
-    }
-
-    const startTime = Date.now()
-    return promise
-      .then((result) => {
-        const time = (Date.now() - startTime) / 1000
-        const noPad = message.length > 40
-        const pad = noPad ? '' : Array(40 - message.length)
-          .fill('-')
-          .join('')
-        t.pass(`${message} ${pad} ${time}s`)
-        return result
-      })
-      .catch((err) => {
-        console.log('error', err);
-        t.error(err, message)
-        if (keepThrowing) throw err
-      })
-  }
-}
 
 const SLEEP_DONE = '▓'
 const SLEEP_TODO = '░'
 
-export function sleep(durationInMs: number) {
+export function sleep (durationInMs: number) {
   return new Promise((resolve) => {
     let count = 0
 
@@ -109,6 +58,3 @@ function undoLastLine() {
   readline.cursorTo(process.stdout, 4) // indent
 }
 
-export function makeSeed () {
-  return '0x' + randomBytes(32).toString('hex')
-}
